requirements.txt
langchain
langsmith
pandas
pyarrow
docker
swebench
python-dotenv
transformers
torch
openai